Prevent Safari caching presign requests
Apparently Safari would sometimes cache the presign GET requests, so
when multiple files are chosen at once, this would often result in the
same file being duplicated. We fix this issue by adding a timestamp to
the presign request, which also what $.ajax does when you tell it to
prevent caching.