Set paramName explicitly to "file" for S3
Often people will have the file field as part of the form, if they're
just uploading single files, or they want to provide a non-JavaScript
fallback. In that case the file field "name" will be the name of the
attachment attribute, however, for Amazon S3 we always need to send the
file under the "file" parameter.